+++ apps/frontend/src/features/admin-form/settings/components/FormEmailSection.tsx
@@ -25,6 +25,7 @@ import FormErrorMessage from '~components/FormControl/FormErrorMessage'
 import FormLabel from '~components/FormControl/FormLabel'
 import { TagInput } from '~components/TagInput'
 
+import { useIsWorkflowBuilderRedesign } from '~features/admin-form/create/workflow/hooks/useIsWorkflowBuilderRedesign'
 import { useUser } from '~features/user/queries'
 
 import { useMutateFormSettings } from '../mutations'
@@ -104,6 +105,7 @@ export const FormEmailSection = ({
   isHighContrast = true,
 }: EmailFormSectionProps): JSX.Element => {
   const { t } = useTranslation()
+  const isRedesign = useIsWorkflowBuilderRedesign()
   const initialEmailSet = useMemo(
     () => new Set(settings.emails),
     [settings.emails],
@@ -160,7 +162,9 @@ export const FormEmailSection = ({
               opacity="1"
             >
               {t(
-                'features.adminForm.settings.emailNotifications.section.regular.description',
+                isRedesign
+                  ? 'features.adminForm.settings.emailNotifications.section.regular.descriptionRedesign'
+                  : 'features.adminForm.settings.emailNotifications.section.regular.description',
               )}
             </FormLabel.Description>
           ) : null}