+++ apps/frontend/src/i18n/locales/features/admin-form/settings/email-notifications/index.ts
@@ -19,6 +19,7 @@ export interface EmailNotifications extends HasTitle {
         stepN: {
           label: {
             overall: string
+            overallRedesign: string
             each: string
           }
           placeholder: string
@@ -26,6 +27,7 @@ export interface EmailNotifications extends HasTitle {
         others: {
           label: string
           description: string
+          descriptionRedesign: string
           tooltipText: string
         }
       }
@@ -33,9 +35,13 @@ export interface EmailNotifications extends HasTitle {
     regular: {
       label: string
       info: string
+      infoRedesign: string
       description: string
+      descriptionRedesign: string
       statusTrackerInfo: string
+      statusTrackerInfoRedesign: string
       statusTrackerDescription: string
+      statusTrackerDescriptionRedesign: string
     }
   }
 }