+++ apps/frontend/src/i18n/locales/features/admin-form/settings/multi-language/en-sg.ts
@@ -1,0 +1,25 @@
+export const enSG = {
+  title: 'Multi-language',
+  toggle: {
+    label: 'Enable multi-language',
+    description:
+      'This will allow respondents to select a language they prefer to view your form in. Translations are not automated.',
+  },
+  languageRow: {
+    hideShowTooltip: 'Hide/show language on form',
+    editTooltip: 'Edit translation',
+    selectDefaultAriaLabel: "Select {language} as the form's default language",
+    addTranslationsAriaLabel: 'Add {language} translations',
+  },
+  sections: {
+    question: 'Question',
+    description: 'Description',
+    column: 'Column',
+    options: 'Options',
+    disableSubmission: 'Disable Submission',
+  },
+  errors: {
+    optionsCountMismatch:
+      'Make sure the number of translated options match the options in the question.',
+  },
+}