+++ packages/sdk/src/crypto-v3.ts
@@ -1,16 +1,15 @@
 import {
-  decodeBase64,
-  decodeUTF8,
-  encodeBase64,
-  encodeUTF8,
-} from 'tweetnacl-util'
-
-import {
   decryptContent,
   encryptMessage,
   generateKeypair,
   verifySignedMessage,
 } from './util/crypto'
+import {
+  decodeBase64,
+  decodeUTF8,
+  encodeBase64,
+  encodeUTF8,
+} from './util/encoding'
 import { determineIsFormFieldsV3 } from './util/validate'
 import { adaptV3ToV4, deriveQuestionFromMeta } from './adapt-v3-to-v4'
 import CryptoBase from './crypto-base'